reinforce_bot/requirements.txt
tensorflow-directml==1.15.5
